// BackEnd/BE_jeework/src/app/_metronic/partials/layout/extras/jee-chat/my-chat/models/message.ts
export  class  Message {
    IdChat: number;
    UserName: string;
    Content_mess: string;
    PathFile: string;
    IsDelAll: boolean;
    IsHidenAll: boolean;
    IsVideoFile: boolean;
    TypeFile: string;
    TenFile: string;
    IdGroup: number;
    Note:string;
    Attachment:any[];
    isInsertMember:boolean;
    CreatedDate: Date;
}